Initial JUnit5 integration for Serenity BDD
Contributors:
  @amohrig
  @DevQuandt
  @fabianlinz
  @sebastianjkl
  @ufroehling
  @worktj

This is actually a squashed version of several commits. Some notes from the original commits:

---
changed package name (junit -> junit5) to avoid clashes in the time it takes for other projects to migrate their tests
---
Remove Junit4 as transitive dependency. JUnit4 is still needed at runtime but at least users of Junit5 should not have it in the classpath.

For gradle this works fine. In Intellij org.junit.Test is still visible and does not result in a compile error in the IDE.
---
Handle assertion errors and exceptions in lifecycle methods (BeforeAll, BeforeEach, AfterEach, AfterAll) (aka "testStarted" field necessary in SerenityJUnitLifecycleAdapterExtension)

## Original question: Do we need a testStarted flag analog to in JUnitStepListener#testStarted
Flag is used in JUnitStepListener to know if JUnitStepListener#testStarted was called before JUnitStepListener#testFailure is called and if this is not the case mark a scenario
called "Initialisation" as started.  This will happen for exceptions or assertion errors in Junit4 @BeforeClass lifecycle methods. In Junit5 there is an explicit hook for this, so we don't have to guess.
=> testStarted flag not necessary for Junit5

## Scenarios
* exception in @BeforeEach (Junit4: @Before) => as if failed in test method itself
* assertion error in @BeforeEach (Junit4: @Before) => as if failed in test method itself
* AfterEach ...
* runtime exception in @BeforeAll (Junit: @BeforeClass) => Report: one entry as broken (instead of individual test methods) (Scenario name: Initialisation)
* assertion error in  @BeforeAll (Junit: @BeforeClass) => Report: one entry as broken (instead of individual test methods) (Scenario name: Initialisation)
* AfterAll ...
* setup is invalid (e.g. non static method for Junit5 @BeforeAll/Junit4 @BeforeClass) => Nothing reported (not ideal IMHO but consistent with JUnit4 support)
* runtime exception in @BeforeAll (Junit: @BeforeClass) => Report: one entry as broken (instead of individual test methods) (Scenario name: Initialisation)
* assertion error in  @BeforeAll (Junit: @BeforeClass) => Report: one entry as broken (instead of individual test methods) (Scenario name: Initialisation)
* exception in Junit4 rule or Junit5 extension
---
Implement BeforeEachCallback in SerenityJUnitLifecycleAdapterExtension instead of BeforeTestExecutionCallback.
Consider call of StepEventBus#clear in SerenityJUnitLifecycleAdapterExtension#beforeEach sufficient to address StepEventBus#initialiseSession in SerenityRunner#runChild.

## Original question: Junit4 SerenityRunner#runChild calls StepEventBus#initialiseSession. Where should this be done for Junit5?

StepEventBus#initialiseSession under certain conditions calls Serenity#clearCurrentSession. This seems kind of redundant, to
JUnitStepListener#testStarted, which always calls StepEventBus#clear, which under the same condition as StepEventBus#initialiseSession calls
Serenity#clearCurrentSession. So the only thing IMHO relevant here it the "correct"/analog timing: Clear is executed before actual page initialization is done.

## Change of lifecycle method
Why is the JUnit5 module using different lifecycle methods? PageExtension#beforeEach but SerenityJUnitLifecycleAdapterExtension#beforeTestExecution
IMHO there is not reason (to be analog to JUnit4 implementation or in general for JUnit5) to use different lifecycle methods here.
IMHO beforeEach is the better fit as it includes @BeforeEach and other extensions on this level.
---
Handle assumption inside and outside of steps & ensure failures after a failing step are ignored

## Junit4 and Junit5 assumptions work differently
* Junit5 org.junit.jupiter.api.Assumptions throws org.opentest4j.TestAbortedException
* Junit4 org.junit.Assume throws org.junit.AssumptionViolatedException
** AssertJ will determine the type to use based on the classpath. As long as Junit4 is on the classpath a Junit4 AssumptionViolatedException is thrown: org.assertj.core.api.Assumptions#assumptionNotMet(java.lang.AssertionError)
* net.thucydides.core.steps.StepInterceptor#executeTestStepMethod only handles Junit4 AssumptionViolatedException inside a step**
** Failing Junit5 assumption are considered as error for the step. A workaround in SerenityStepExtension addresses this issue
** Currently net.serenitybdd.junit.extension.SerenityStepExtension#checkForStepAssumptionViolations can never actually be executed because the StepInterceptor does not take TestAbortedException into account
---